requirements.txt
Flask==3.1.2
litellm==1.80.15
python-dotenv